GitHub-API-CLI/requirements.txt
requests
python-dotenv
prettytable

